ASNIDE-228 Renamed message to messages
