+++ src/asn1sccserviceprovider.cpp
@@ -34,8 +34,8 @@
 #include <coreplugin/icore.h>
 
 #include <utils/qtcassert.h>
-#include <message/messageutils.h>
-#include <message/messagemanager.h>
+#include <messages/messageutils.h>
+#include <messages/messagemanager.h>
 
 #include "asn1acnconstants.h"
 
@@ -52,7 +52,7 @@ Asn1SccServiceProvider::Asn1SccServiceProvider(Settings::ServiceConstPtr setting
 
     connect(m_asn1sccService, &QProcess::stateChanged,
             [this](QProcess::ProcessState newState) {
-        Q_UNUSED(newState); Message::messageProcess(m_asn1sccService);
+        Q_UNUSED(newState); Messages::messageProcess(m_asn1sccService);
     } );
 
     connect(m_settings.get(), &Settings::Service::changed, this, &Asn1SccServiceProvider::settingsChanged);
@@ -84,28 +84,28 @@ QNetworkReply *Asn1SccServiceProvider::requestAst(const QHash<QString, QString> 
 void Asn1SccServiceProvider::start()
 {
     QTC_ASSERT(m_serviceStarted == false, return);
-    Message::MessageManager::write("Starting Asn1Scc service...",
-                                   Message::MessageManager::Type::Process);
+    Messages::MessageManager::write("Starting Asn1Scc service...",
+                                    Messages::MessageManager::Type::Process);
     m_serviceStarted = true;
 
     m_stayAliveTimer.start();
     m_asn1sccService->start();
 
     if (!m_asn1sccService->waitForStarted()) // TODO handle error?
-        Message::messageProcessError(m_asn1sccService);
+        Messages::messageProcessError(m_asn1sccService);
 }
 
 void Asn1SccServiceProvider::stop()
 {
-    Message::MessageManager::write("Stopping Asn1Scc service...",
-                                   Message::MessageManager::Type::Process);
+    Messages::MessageManager::write("Stopping Asn1Scc service...",
+                                    Messages::MessageManager::Type::Process);
     m_serviceStarted = false;
 
     m_stayAliveTimer.stop();
     m_asn1sccService->terminate();
 
     if (!m_asn1sccService->waitForFinished()) // TODO handle error?
-        Message::messageProcessError(m_asn1sccService);
+        Messages::messageProcessError(m_asn1sccService);
 }
 
 QStringList Asn1SccServiceProvider::additionalArguments() const
@@ -147,8 +147,8 @@ void Asn1SccServiceProvider::onProcessFinished(int exitCode, QProcess::ExitStatu
     Q_UNUSED(exitCode);
     Q_UNUSED(exitStatus);
 
-    Message::MessageManager::write("Process finished...",
-                                   Message::MessageManager::Type::Process);
+    Messages::MessageManager::write("Process finished...",
+                                    Messages::MessageManager::Type::Process);
 
     if (m_serviceStarted) {
         m_asn1sccService->start();
@@ -169,7 +169,7 @@ void Asn1SccServiceProvider::onStayAliveRequestFinished()
     QTC_ASSERT(reply != nullptr, return);
 
     if (m_serviceStarted && reply->error() != QNetworkReply::NoError) {
-        Message::messageNetworkReplyError(reply);
+        Messages::messageNetworkReplyError(reply);
         restart();
     }
